feat(engine): live query の値を u64 に (64 bit 列の購読・集計・上位 k 件) (段階 D)
live.rs の値の型を u64 に広げる (entity と鍵 id は u32 のまま)。 段階 A で入れた 「live は 64 bit 列を断る」
入口を撤去し、 64 bit 列を条件 / group / 合計 / 並びに使えるようにする。

- LivePred::{Eq, Range, In} の値、 CellReader (cell / pull / pull_range / pull_len) の値を u64。 ref をたどる所は
  ref_cell (Ref の列は local eid = u32)
- 条件 / 穴の値 / 形の符号 / 鍵の組 / 鍵の表 / 範囲の帯 (値の最大は u64::MAX - 1) / 区間の索引 / 根の答え /
  上位 k 件の境界と順序索引 ((値, eid)) / group の値を u64、 合計 (Agg.sum) は u128
- 値のページ (範囲 / group / 並びの値、 合計する値) は ValWords: u32 に収まる間は 4 B、 収まらない値が来たら
  8 B に作り直す (u32 の列だけの購読の状態は大きくならない)
- LiveCounts の group の値を u64 に。 schema の wrapper は u32 の列の値に戻し、 合計は u64 に (飽和)
- subscribe_sums は Number64 の列の合計も受ける

下流から見た破壊的変更: LivePred の値が u64 (u32 の変数は `.into()`)、 LiveCounts::poll / poll_sums / get /
all の group が u64、 Agg.sum が u128。

テスト: 64 bit 列の購読の oracle (等値 / 範囲 / In / 否定 / ref の先の範囲 / 範囲 2 本 / 大きな値が入った後に
張る範囲 (ref の先を含む) / 64 bit 列で group の件数 / 64 bit の値の合計 / 上位 k 件 昇順・降順・ref の先)、
小さな値で状態を作ってから大きな値を混ぜる (ページが 8 B に作り直される)。 最後の帯の古い記録を忘れさせる
単体テスト。 変異 7 本で落ちる。 既存の live のテストはそのまま通る。 workspace 1213 passed

実測 (u32 の列、 ivm-bench、 旧 = 段階 C と交互に): flip p50 3.21-3.88 → 3.25-3.88 µs、 repoint
0.38-0.46 → 0.38-0.46 µs、 hacg 会社単位 move 0.25 → 0.25 µs、 hacg 集計 move 1.04 → 1.08-1.17 µs、 差分の
件数は全部一致。 上位 k 件の順序索引は 1 件 8 → 16 B

+++ crates/enchudb-engine/tests/live_query.rs
@@ -476,8 +476,8 @@ fn new_kinds_under_concurrent_writes() {
             .unwrap();
         let qs: [&LiveQuery; 6] = [&or, &range, &range_via, &top, &top_via, &top_in];
         let mut seen: Vec<BTreeSet<u64>> = vec![BTreeSet::new(); qs.len()];
-        let mut groups: std::collections::BTreeMap<u32, u64> = Default::default();
-        let mut sum_groups: std::collections::BTreeMap<u32, enchudb_engine::Agg> = Default::default();
+        let mut groups: std::collections::BTreeMap<u64, u64> = Default::default();
+        let mut sum_groups: std::collections::BTreeMap<u64, enchudb_engine::Agg> = Default::default();
         let group = eng.live_group();
         for q in qs {
             group.add(q);
@@ -549,20 +549,20 @@ fn new_kinds_under_concurrent_writes() {
             assert_eq!(seen[i], want[i], "round {round}: [{}] 積分 != 手で数えた結果", names[i]);
             assert_eq!(qs[i].count(&eng), want[i].len(), "round {round}: [{}] count", names[i]);
         }
-        let mut want_groups: std::collections::BTreeMap<u32, u64> = Default::default();
+        let mut want_groups: std::collections::BTreeMap<u64, u64> = Default::default();
         for &e in &users {
             if get(&eng, e, "age").is_some() && let Some(c) = city_of(e) {
-                *want_groups.entry(c).or_insert(0) += 1;
+                *want_groups.entry(c as u64).or_insert(0) += 1;
             }
         }
         assert_eq!(groups, want_groups, "round {round}: [counts] 積分 != 手で数えた件数");
         assert_eq!(counts.all(&eng), want_groups.into_iter().collect::<Vec<_>>(), "round {round}: [counts] all");
-        let mut want_sums: std::collections::BTreeMap<u32, enchudb_engine::Agg> = Default::default();
+        let mut want_sums: std::collections::BTreeMap<u64, enchudb_engine::Agg> = Default::default();
         for &e in &users {
             if get(&eng, e, "age").is_some_and(|a| a <= 25) && let Some(c) = city_of(e) {
-                let w = want_sums.entry(c).or_default();
+                let w = want_sums.entry(c as u64).or_default();
                 w.count += 1;
-                w.sum += get(&eng, e, "score").unwrap_or(0) as u64;
+                w.sum += get(&eng, e, "score").unwrap_or(0) as u128;
             }
         }
         assert_eq!(sum_groups, want_sums, "round {round}: [sums] 積分 != 手で数えた件数 / 合計");
